+++ libasynql/src/poggit/libasynql/ConfigException.php
@@ -20,29 +20,12 @@
 
 declare(strict_types=1);
 
-namespace poggit\libasynql\mysql;
+namespace poggit\libasynql;
 
-use poggit\libasynql\result\SqlColumnInfo;
+use InvalidArgumentException;
 
-class MysqlColumnInfo extends SqlColumnInfo{
-	private $flags;
-	private $mysqlType;
-
-	public function __construct(string $name, string $type, int $flags, int $mysqlType){
-		parent::__construct($name, $type);
-		$this->flags = $flags;
-		$this->mysqlType = $mysqlType;
-	}
-
-	public function getFlags() : int{
-		return $this->flags;
-	}
-
-	public function hasFlag(int $flag){
-		return ($this->flags & $flag) > 0;
-	}
-
-	public function getMysqlType() : int{
-		return $this->mysqlType;
+class ConfigException extends InvalidArgumentException{
+	public function __construct(string $message = ""){
+		parent::__construct("Config problem: " . $message);
 	}
 }