+++ libasynql/src/poggit/libasynql/DataConnector.php
@@ -22,16 +22,22 @@ declare(strict_types=1);
 
 namespace poggit\libasynql;
 
+use InvalidArgumentException;
+use poggit\libasynql\generic\GenericStatementFileParseException;
+
 /**
  * Represents a database connection or a group of database connections
  */
-interface DatabaseConnection{
+interface DataConnector{
 	/**
 	 * Loads pre-formatted queries from a readable stream resource.
 	 *
 	 * The implementation will close the stream after reading.
 	 *
 	 * @param resource $fh a stream that supports <code>feof()</code>, <code>fgets()</code> and <code>fclose()</code>.
+	 *
+	 * @throws GenericStatementFileParseException if the file contains a syntax error or compile error
+	 * @throws InvalidArgumentException if the file introduces statements that duplicate the names of those previously loaded
 	 */
 	public function loadQueryFile($fh) : void;
 
@@ -39,48 +45,49 @@ interface DatabaseConnection{
 	 * Loads a pre-formatted query.
 	 *
 	 * @param GenericStatement $stmt
+	 * @throws InvalidArgumentException if the statement duplicates the name of one previously loaded
 	 */
 	public function loadQuery(GenericStatement $stmt) : void;
 
 	/**
 	 * Executes a generic query that either succeeds or fails.
 	 *
-	 * @param string        $name      the {@link GenericPreparedStatement} query name
+	 * @param string        $queryName the {@link GenericPreparedStatement} query name
 	 * @param mixed[]       $args      the variables as defined in the {@link GenericPreparedStatement}
 	 * @param callable|null $onSuccess an optional callback when the query has succeeded: <code>function() : void{}</code>
 	 * @param callable|null $onError   an optional callback when the query has failed: <code>function({@link SqlError} $error) : void{}</code>
 	 */
-	public function executeGeneric(string $name, array $args, ?callable $onSuccess = null, ?callable $onError = null) : void;
+	public function executeGeneric(string $queryName, array $args, ?callable $onSuccess = null, ?callable $onError = null) : void;
 
 	/**
 	 * Executes a query that changes data.
 	 *
-	 * @param string        $name      the {@link GenericPreparedStatement} query name
+	 * @param string        $queryName the {@link GenericPreparedStatement} query name
 	 * @param mixed[]       $args      the variables as defined in the {@link GenericPreparedStatement}
 	 * @param callable|null $onSuccess an optional callback when the query has succeeded: <code>function(int $affectedRows) : void{}</code>
 	 * @param callable|null $onError   an optional callback when the query has failed: <code>function({@link SqlError} $error) : void{}</code>
 	 */
-	public function executeChange(string $name, array $args, ?callable $onSuccess = null, ?callable $onError = null) : void;
+	public function executeChange(string $queryName, array $args, ?callable $onSuccess = null, ?callable $onError = null) : void;
 
 	/**
 	 * Executes an insert query that results in an insert ID.
 	 *
-	 * @param string        $name       the {@link GenericPreparedStatement} query name
+	 * @param string        $queryName  the {@link GenericPreparedStatement} query name
 	 * @param mixed[]       $args       the variables as defined in the {@link GenericPreparedStatement}
-	 * @param callable|null $onInserted an optional callback when the query has succeeded: <code>function(int $insertId) : void{}</code>
+	 * @param callable|null $onInserted an optional callback when the query has succeeded: <code>function(int $insertId, callable $affectedRows) : void{}</code>
 	 * @param callable|null $onError    an optional callback when the query has failed: <code>function({@link SqlError} $error) : void{}</code>
 	 */
-	public function executeInsert(string $name, array $args, ?callable $onInserted = null, ?callable $onError = null) : void;
+	public function executeInsert(string $queryName, array $args, ?callable $onInserted = null, ?callable $onError = null) : void;
 
 	/**
 	 * Executes a select query that returns an SQL result set. This does not strictly need to be SELECT queries -- reflection queries like MySQL's <code>SHOW TABLES</code> query are also allowed.
 	 *
-	 * @param string        $name     the {@link GenericPreparedStatement} query name
-	 * @param mixed[]       $args     the variables as defined in the {@link GenericPreparedStatement}
-	 * @param callable|null $onSelect an optional callback when the query has succeeded: <code>function(int $insertId) : void{}</code>
-	 * @param callable|null $onError  an optional callback when the query has failed: <code>function({@link SqlError} $error) : void{}</code>
+	 * @param string        $queryName the {@link GenericPreparedStatement} query name
+	 * @param mixed[]       $args      the variables as defined in the {@link GenericPreparedStatement}
+	 * @param callable|null $onSelect  an optional callback when the query has succeeded: <code>function({@link \poggit\libasynql\result\SqlSelectResult SqlSelectResult} $result) : void{}</code>
+	 * @param callable|null $onError   an optional callback when the query has failed: <code>function({@link SqlError} $error) : void{}</code>
 	 */
-	public function executeSelect(string $name, array $args, ?callable $onSelect = null, ?callable $onError = null) : void;
+	public function executeSelect(string $queryName, array $args, ?callable $onSelect = null, ?callable $onError = null) : void;
 
 	/**
 	 * Closes the connection and/or all child connections. Remember to call this method when the plugin is disabled or the data provider is switched.